Extract settings dialog to ui/settings_dialog.py; drop dead aliases and imports
